# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sankardev Vidya Niketan School API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login (staff or student)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current session user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Not authenticated
...
  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List news and announcements
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: News articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsItem"
    post:
      operationId: createNews
      tags: [news]
      summary: Create news/announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"
  /news/{id}:
    patch:
      operationId: updateNews
      tags: [news]
      summary: Update a news item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"
    delete:
      operationId: deleteNews
      tags: [news]
      summary: Delete a news item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admissions:
    get:
      operationId: listAdmissions
      tags: [admission]
      summary: List admission applications
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: Admission applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdmissionApplication"
    post:
      operationId: submitAdmission
      tags: [admission]
      summary: Submit an admission application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdmissionInput"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdmissionApplication"
  /admissions/{id}:
    get:
      operationId: getAdmission
      tags: [admission]
      summary: Get an admission application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Admission application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdmissionApplication"
    patch:
      operationId: updateAdmissionStatus
      tags: [admission]
      summary: Update admission status (approve/reject)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdmissionStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdmissionApplication"
...
  /attendance:
    get:
      operationId: getAttendance
      tags: [attendance]
      summary: Get attendance records
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: className
          in: query
          schema:
            type: string
        - name: section
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
  /attendance/bulk:
    post:
      operationId: markAttendanceBulk
      tags: [attendance]
      summary: Mark attendance for a class on a date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceBulkInput"
      responses:
        "200":
          description: Marked
          content:
            application/json:
              schema:
                type: object
                required: [ok, count]
                properties:
                  ok:
                    type: boolean
                  count:
                    type: integer
  /attendance/report:
    get:
      operationId: getAttendanceReport
      tags: [attendance]
      summary: Get monthly attendance report
      parameters:
        - name: className
          in: query
          required: true
          schema:
            type: string
        - name: section
          in: query
          schema:
            type: string
        - name: month
          in: query
          required: true
          schema:
            type: string
        - name: year
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Monthly attendance report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceReport"
  /attendance/{id}:
    delete:
      operationId: deleteAttendance
      tags: [attendance]
      summary: Delete an attendance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password, role]
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, staff, student]
    AuthResponse:
      type: object
      required: [id, username, role, name]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
        name:
          type: string
...
    NewsItem:
      type: object
      required: [id, title, content, publishedAt, isImportant]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        titleAssamese:
          type: ["string", "null"]
        contentAssamese:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        isImportant:
          type: boolean
        publishedAt:
          type: string
    NewsInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
        isImportant:
          type: boolean
    NewsUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
        isImportant:
          type: boolean
    AdmissionApplication:
      type: object
      required: [id, serialNumber, studentName, fatherName, motherName, appliedForClass, status, submittedAt]
      properties:
        id:
          type: integer
        serialNumber:
          type: ["string", "null"]
        classRollNumber:
          type: ["string", "null"]
        studentName:
          type: string
        fatherName:
          type: string
        motherName:
          type: string
        permanentAddress:
          type: ["string", "null"]
        presentAddress:
          type: ["string", "null"]
        pinCode:
          type: ["string", "null"]
        guardianName:
          type: ["string", "null"]
        guardianRelation:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        age:
          type: ["string", "null"]
        caste:
          type: ["string", "null"]
        religion:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        bloodGroup:
          type: ["string", "null"]
        previousSchoolName:
          type: ["string", "null"]
        previousSchoolAddress:
          type: ["string", "null"]
        previousClass:
          type: ["string", "null"]
        previousClassResult:
          type: ["string", "null"]
        reasonForLeaving:
          type: ["string", "null"]
        appliedForClass:
          type: string
        apaarId:
          type: ["string", "null"]
        siblingName:
          type: ["string", "null"]
        siblingClass:
          type: ["string", "null"]
        siblingSection:
          type: ["string", "null"]
        specialCategory:
          type: ["string", "null"]
        fatherPhone:
          type: ["string", "null"]
        motherPhone:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        submittedAt:
          type: string
        remarks:
          type: ["string", "null"]
    AdmissionInput:
      type: object
      required: [studentName, fatherName, motherName, appliedForClass]
      properties:
        serialNumber:
          type: string
        classRollNumber:
          type: string
        studentName:
          type: string
        fatherName:
          type: string
        motherName:
          type: string
        permanentAddress:
          type: string
        presentAddress:
          type: string
        pinCode:
          type: string
        guardianName:
          type: string
        guardianRelation:
          type: string
        dateOfBirth:
          type: string
        age:
          type: string
        caste:
          type: string
        religion:
          type: string
        nationality:
          type: string
        bloodGroup:
          type: string
        previousSchoolName:
          type: string
        previousSchoolAddress:
          type: string
        previousClass:
          type: string
        previousClassResult:
          type: string
        reasonForLeaving:
          type: string
        appliedForClass:
          type: string
        apaarId:
          type: string
        siblingName:
          type: string
        siblingClass:
          type: string
        siblingSection:
          type: string
        specialCategory:
          type: string
        fatherPhone:
          type: string
        motherPhone:
          type: string
    AdmissionStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, approved, rejected]
        remarks:
          type: string
...
    AttendanceRecord:
      type: object
      required: [id, studentId, date, status, className]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        rollNumber:
          type: ["string", "null"]
        date:
          type: string
        status:
          type: string
          enum: [present, absent, late]
        className:
          type: string
        section:
          type: ["string", "null"]
        markedBy:
          type: ["string", "null"]
        remarks:
          type: ["string", "null"]
    AttendanceBulkInput:
      type: object
      required: [date, className, records]
      properties:
        date:
          type: string
        className:
          type: string
        section:
          type: string
        markedBy:
          type: string
        records:
          type: array
          items:
            type: object
            required: [studentId, status]
            properties:
              studentId:
                type: integer
              status:
                type: string
                enum: [present, absent, late]
              remarks:
                type: string
    AttendanceReport:
      type: object
      required: [className, month, year, startDate, endDate, students]
      properties:
        className:
          type: string
        section:
          type: ["string", "null"]
        month:
          type: integer
        year:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        students:
          type: array
          items:
            type: object
            required: [studentId, studentName, rollNumber, present, absent, late, total, dates]
            properties:
              studentId:
                type: integer
              studentName:
                type: string
              rollNumber:
                type: string
              present:
                type: integer
              absent:
                type: integer
              late:
                type: integer
              total:
                type: integer
              dates:
                type: object
                additionalProperties:
                  type: string
...